Move HashedAddressMapping into test
